follow camera into reactor
