+++ packages/spatial/src/camera/components/FollowCameraComponent.ts
@@ -23,14 +23,31 @@ All portions of the code written by the Ethereal Engine team are Copyright © 20
 Ethereal Engine. All Rights Reserved.
 */
 
-import { ArrowHelper, Clock, Vector3 } from 'three'
-
+import { ECSState, Engine, defineQuery, useEntityContext } from '@etherealengine/ecs'
+import {
+  defineComponent,
+  getComponent,
+  getOptionalComponent,
+  removeComponent
+} from '@etherealengine/ecs/src/ComponentFunctions'
+import { Entity, UndefinedEntity } from '@etherealengine/ecs/src/Entity'
 import { getState } from '@etherealengine/hyperflux'
-
-import { defineComponent } from '@etherealengine/ecs/src/ComponentFunctions'
-import { UndefinedEntity } from '@etherealengine/ecs/src/Entity'
+import { useEffect } from 'react'
+import { ArrowHelper, Clock, MathUtils, Matrix4, Raycaster, Vector3 } from 'three'
+import { TransformComponent } from '../../SpatialModule'
+import { createConeOfVectors } from '../../common/functions/MathFunctions'
+import { smoothDamp } from '../../common/functions/MathLerpFunctions'
+import { MeshComponent } from '../../renderer/components/MeshComponent'
+import { ObjectLayerComponents } from '../../renderer/components/ObjectLayerComponent'
+import { VisibleComponent } from '../../renderer/components/VisibleComponent'
+import { ObjectLayers } from '../../renderer/constants/ObjectLayers'
+import {
+  ComputedTransformComponent,
+  setComputedTransformComponent
+} from '../../transform/components/ComputedTransformComponent'
 import { CameraSettingsState } from '../CameraSceneMetadata'
 import { CameraMode } from '../types/CameraMode'
+import { TargetCameraRotationComponent } from './TargetCameraRotationComponent'
 
 export const coneDebugHelpers: ArrowHelper[] = []
 
@@ -112,5 +129,158 @@ export const FollowCameraComponent = defineComponent({
     if (typeof json.maxPhi !== 'undefined') component.maxPhi.set(json.maxPhi)
     if (typeof json.shoulderSide !== 'undefined') component.shoulderSide.set(json.shoulderSide)
     if (typeof json.locked !== 'undefined') component.locked.set(json.locked)
+  },
+
+  reactor: () => {
+    const entity = useEntityContext()
+
+    useEffect(() => {
+      const followCamera = getComponent(entity, FollowCameraComponent)
+      setComputedTransformComponent(entity, followCamera.targetEntity, computeCameraFollow)
+
+      return () => {
+        removeComponent(entity, ComputedTransformComponent)
+      }
+    }, [])
+
+    return null
   }
 })
+
+const targetPosition = new Vector3()
+const direction = new Vector3()
+const upVector = new Vector3(0, 1, 0)
+const empty = new Vector3()
+const mx = new Matrix4()
+const tempVec1 = new Vector3()
+const raycaster = new Raycaster()
+
+const computeCameraFollow = (cameraEntity: Entity, referenceEntity: Entity) => {
+  const followCamera = getComponent(cameraEntity, FollowCameraComponent)
+  const cameraTransform = getComponent(cameraEntity, TransformComponent)
+  const targetTransform = getComponent(referenceEntity, TransformComponent)
+
+  if (!targetTransform || !followCamera) return
+
+  // Limit the pitch
+  followCamera.phi = Math.min(followCamera.maxPhi, Math.max(followCamera.minPhi, followCamera.phi))
+
+  let maxDistance = followCamera.zoomLevel
+  let isInsideWall = false
+
+  targetPosition
+    .copy(followCamera.offset)
+    .applyQuaternion(TransformComponent.getWorldRotation(referenceEntity, targetTransform.rotation))
+    .add(TransformComponent.getWorldPosition(referenceEntity, new Vector3()))
+
+  // Run only if not in first person mode
+  if (followCamera.raycastProps.enabled && followCamera.zoomLevel >= followCamera.minDistance) {
+    const distanceResults = getMaxCamDistance(cameraEntity, targetPosition)
+    maxDistance = distanceResults.maxDistance
+    isInsideWall = distanceResults.targetHit
+  }
+
+  const newZoomDistance = Math.min(followCamera.zoomLevel, maxDistance)
+
+  // Zoom smoothing
+  const smoothingSpeed = isInsideWall ? 0.1 : 0.3
+  const deltaSeconds = getState(ECSState).deltaSeconds
+
+  followCamera.distance = smoothDamp(
+    followCamera.distance,
+    newZoomDistance,
+    followCamera.zoomVelocity,
+    smoothingSpeed,
+    deltaSeconds
+  )
+
+  const theta = followCamera.theta
+  const thetaRad = MathUtils.degToRad(theta)
+  const phiRad = MathUtils.degToRad(followCamera.phi)
+
+  cameraTransform.position.set(
+    targetPosition.x + followCamera.distance * Math.sin(thetaRad) * Math.cos(phiRad),
+    targetPosition.y + followCamera.distance * Math.sin(phiRad),
+    targetPosition.z + followCamera.distance * Math.cos(thetaRad) * Math.cos(phiRad)
+  )
+
+  direction.copy(cameraTransform.position).sub(targetPosition).normalize()
+
+  mx.lookAt(direction, empty, upVector)
+  cameraTransform.rotation.setFromRotationMatrix(mx)
+
+  updateCameraTargetRotation(cameraEntity)
+}
+
+const updateCameraTargetRotation = (cameraEntity: Entity) => {
+  if (!cameraEntity) return
+  const followCamera = getComponent(cameraEntity, FollowCameraComponent)
+  const target = getOptionalComponent(cameraEntity, TargetCameraRotationComponent)
+  if (!target) return
+
+  const epsilon = 0.001
+
+  target.phi = Math.min(followCamera.maxPhi, Math.max(followCamera.minPhi, target.phi))
+
+  if (Math.abs(target.phi - followCamera.phi) < epsilon && Math.abs(target.theta - followCamera.theta) < epsilon) {
+    removeComponent(followCamera.targetEntity, TargetCameraRotationComponent)
+    return
+  }
+
+  const delta = getState(ECSState).deltaSeconds
+  followCamera.phi = smoothDamp(followCamera.phi, target.phi, target.phiVelocity, target.time, delta)
+  followCamera.theta = smoothDamp(followCamera.theta, target.theta, target.thetaVelocity, target.time, delta)
+}
+
+const cameraLayerQuery = defineQuery([VisibleComponent, ObjectLayerComponents[ObjectLayers.Camera], MeshComponent])
+
+const getMaxCamDistance = (cameraEntity: Entity, target: Vector3) => {
+  const followCamera = getComponent(cameraEntity, FollowCameraComponent)
+
+  // Cache the raycast result for 0.1 seconds
+  const raycastProps = followCamera.raycastProps
+  const { camRayCastCache, camRayCastClock, cameraRays, rayConeAngle } = raycastProps
+  if (camRayCastCache.maxDistance != -1 && camRayCastClock.getElapsedTime() < raycastProps.rayFrequency) {
+    return camRayCastCache
+  }
+
+  camRayCastClock.start()
+
+  const sceneObjects = cameraLayerQuery().flatMap((e) => getComponent(e, MeshComponent))
+
+  // Raycast to keep the line of sight with avatar
+  const cameraTransform = getComponent(Engine.instance.cameraEntity, TransformComponent)
+  const targetToCamVec = tempVec1.subVectors(cameraTransform.position, target)
+  // raycaster.ray.origin.sub(targetToCamVec.multiplyScalar(0.1)) // move origin behind camera
+
+  createConeOfVectors(targetToCamVec, cameraRays, rayConeAngle)
+
+  let maxDistance = Math.min(followCamera.maxDistance, raycastProps.rayLength)
+
+  // Check hit with mid ray
+  raycaster.layers.set(ObjectLayers.Camera) // Ignore avatars
+  // @ts-ignore - todo figure out why typescript freaks out at this
+  raycaster.firstHitOnly = true // three-mesh-bvh setting
+  raycaster.far = followCamera.maxDistance
+  raycaster.set(target, targetToCamVec.normalize())
+  const hits = raycaster.intersectObjects(sceneObjects, false)
+
+  if (hits[0] && hits[0].distance < maxDistance) {
+    maxDistance = hits[0].distance
+  }
+
+  //Check the cone for minimum distance
+  cameraRays.forEach((rayDir, i) => {
+    raycaster.set(target, rayDir)
+    const hits = raycaster.intersectObjects(sceneObjects, false)
+
+    if (hits[0] && hits[0].distance < maxDistance) {
+      maxDistance = hits[0].distance
+    }
+  })
+
+  camRayCastCache.maxDistance = maxDistance
+  camRayCastCache.targetHit = !!hits[0]
+
+  return camRayCastCache
+}
